Move some main-process files around
