+++ src/main-process/modules/config.js
