+++ src/main-process/modules/power-monitor.js
